Day 2 & its the first commit
messy spaghetti code, it works, just not the way we want.
Lots of work to be done here..
the end ponits are messed up, they are threaded but need to
support dynamic/progressive response so that results dont just pop out
and are more gradual.
need to fill in the config a bit more
did not add db, there is no storing of data as of now. its all real time
we got some shit together as of now i would say

:)